# src/main/java/com/genfood/foodgenback/endpoint/rest/model/Auth.java
package com.genfood.foodgenback.endpoint.rest.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Builder
@Data
public class Auth {
  private String email;
  private String password;
}